Add booking
Add routesBooking, functionForBooking and template

+++ app/model.py
@@ -34,7 +34,7 @@ movies = Table('movies', metadata,
              Column('id', Integer, primary_key = True, autoincrement=True), 
              Column('title', String),#not null
              Column('minimumAge', Integer, default = 0),#CHECK >=0
-             Column('duration', Float, nullable = False),#CHECK >=0
+             Column('duration', Integer, nullable = False),#CHECK >=0
              Column('idGenre', None, ForeignKey('genres.id'))
 
             )
